devenv: split out main library functions
