+++ devenv/src/lib.rs
@@ -1,2 +1,7 @@
-pub mod cli;
+mod cli;
+pub mod command;
+pub mod config;
+pub mod devenv;
 pub mod log;
+
+pub use cli::GlobalOptions;
